feat(i): adding pages
