+++ app/globals.css
@@ -11,6 +11,8 @@
   width: 400px;
   box-shadow: 0 8px 16px rgba(0, 0, 0, 0.2);
   transition: transform 0.3s;
+  margin-top: 60px;
+  margin-left: 2px;
 }
 
 .card-un:hover {
@@ -70,7 +72,7 @@
 .column {
   float: left;
   width: 33.33%;
-  margin-top: 50px;
+  margin-top: 80px;
 }
 
 .ppl-prayed {
