+++ app/home/page.tsx
@@ -3,7 +3,7 @@ import React from "react";
 interface Props {}
 
 const Home = (props: Props) => {
-  return <div>Home</div>;
+  return <div></div>;
 };
 
 export default Home;
