+++ app/page.tsx
@@ -1,31 +1,7 @@
 import React from "react";
 import Links from "./Routes/Routes";
-import Card from "./home/card";
 import Comments from "./home/comments";
 import Home from "./home/page";
-const mockData = [
-  {
-    id: 1,
-    prayer: "I want to be a better person",
-    username: "Alejandro Chang",
-    date: "2021-07-02",
-    likes: 0,
-  },
-  {
-    id: 1,
-    prayer: "I want to be a better person",
-    username: "Juan Espinoza",
-    date: "2021-07-02",
-    likes: 0,
-  },
-  {
-    id: 1,
-    prayer: "I want to be a better person",
-    username: "Filipe",
-    date: "2021-07-02",
-    likes: 0,
-  },
-];
 
 export default function App() {
   /*
@@ -37,10 +13,7 @@ export default function App() {
       <Home />
       <div className="row">
         <div className="column">
-          <Card data={mockData} />
-        </div>
-        <div className="column">
-          <Comments />å
+          <Comments />
         </div>
       </div>
     </div>
